+++ docs/grid-system/src/App.js
@@ -4,10 +4,10 @@ import { Row, Col } from 'react-configurable-grid'
 function App() {
   return (
     <div className="app">
-      <Row>
+      <Row justify="flex-end">
         {
-          [1, 2, 3, 4].map((x, index) => (
-            <Col xs={3} key={index} className="outer-col">
+          [1, 2, 3, 4, 5].map((x, index) => (
+            <Col xs={3} mb={index + 'px'} key={index} className="outer-col">
               <div className="inner-col">col-xs-{index}</div>
             </Col>
           ))