+++ src/Col/index.js
@@ -1,40 +1,35 @@
 import React from 'react';
 import { oneOfType, number, string, node } from "prop-types";
 
-import styles, { colClasses } from "./styles";
+import gerenateColStyles from "./styles";
+import { appendStyles } from '../utils/styleUtils';
 
-function Col({ xs, sm, md, lg, xl, mt, mr, mb, ml, className, children }) {
+function Col({ xs, sm, md, lg, xl, className, children, ...props }) {
   const classes = [
+    'rcg-col',
     className
   ];
 
   if (xs) {
-    classes.push(`col-xs-${xs}`);
+    classes.push(`rcg-col-xs-${xs}`);
   }
   if (sm) {
-    classes.push(`col-sm-${sm}`);
+    classes.push(`rcg-col-sm-${sm}`);
   }
   if (md) {
-    classes.push(`col-md-${md}`);
+    classes.push(`rcg-col-md-${md}`);
   }
   if (lg) {
-    classes.push(`col-lg-${lg}`);
+    classes.push(`rcg-col-lg-${lg}`);
   }
   if (xl) {
-    classes.push(`col-lg-${xl}`);
+    classes.push(`rcg-col-lg-${xl}`);
   }
 
+  appendStyles('column', gerenateColStyles());
+
   return (
-    <div className={classes.join(" ")}>
-      <style jsx>{styles}</style>
-      <style jsx>{`
-        div {
-          margin: ${mt} ${mr} ${mb} ${ml};
-        }
-      `}</style>
-      <style jsx global>
-        {colClasses}
-      </style>
+    <div className={classes.join(" ")} {...props}>
       {children}
     </div>
   );
@@ -46,10 +41,6 @@ Col.defaultProps = {
   md: null,
   lg: null,
   xl: null,
-  mt: 0,
-  mr: 0,
-  mb: 0,
-  ml: 0,
   className: null
 };
 
@@ -59,10 +50,6 @@ Col.propTypes = {
   md: oneOfType([number, string]),
   lg: oneOfType([number, string]),
   xl: oneOfType([number, string]),
-  mt: oneOfType([number, string]),
-  mr: oneOfType([number, string]),
-  mb: oneOfType([number, string]),
-  ml: oneOfType([number, string]),
   className: string,
   children: node.isRequired
 };