+++ src/Row/styles.js
@@ -1,9 +1,55 @@
-import css from "styled-jsx/css";
-import { gutter } from "../constants";
+import { halfGutter } from "../config";
+import formatStyles from "../utils/formatStyles";
 
-export default css`
-  div {
-    display: flex;
-    margin: 0 -${gutter};
-  }
-`;
+const wrapTypes = [
+  'nowrap',
+  'wrap',
+  'wrap-reverse'
+];
+
+const alignTypes = [
+  'baseline',
+  'center',
+  'flex-end',
+  'flex-start',
+  'stretch'
+];
+
+const justifyTypes = alignTypes;
+
+export default function gerenateRowStyles() {
+  let styles = `
+    .rcg-row {
+      display: -webkit-box;
+      display: -webkit-flex;
+      display: -ms-flexbox;
+      display: flex;
+      margin: 0_-${halfGutter};
+    }
+
+    ${wrapTypes.map(wrapType => `
+      .rcg-row-wrap--${wrapType} {
+        -ms-flex-wrap: ${wrapType};
+        flex-wrap: ${wrapType};
+      }
+    `)}
+
+    ${alignTypes.map(alignType => `
+      .rcg-row-align--${alignType} {
+        -webkit-box-align: ${alignType};
+        -ms-flex-align: ${alignType};
+        align-items: ${alignType};
+      }
+    `)}
+
+    ${justifyTypes.map(justifyType => `
+      .rcg-row-justify--${justifyType} {
+        -webkit-box-pack: ${justifyType};
+        -ms-flex-pack: ${justifyType};
+        justify-content: ${justifyType};
+      }
+    `)}
+  `;
+
+  return formatStyles(styles);
+}