+++ src/utils/formatStyles.js
@@ -1,0 +1,6 @@
+
+export default function cleanStyles(styles) {
+  return styles
+    .replace(/\n| |,/g, '')
+    .replace(/_/g, ' ');
+}