+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.0.2 (March 21, 2018)
+
+- Fix missing module
+
+## 2.0.0 (March 21, 2018)
+
+- Removed styled-jx and stylis dependencies
+- *Removed mt, ml, mb and mr props for margin control*
+
 ## 1.1.11 (March 21, 2018)
 
 - Fix errors with styled-jx and stylis